docs/requirements.txt
feedparser==6.0.11
pillow
